[unittests] Factor out TestContext helper for AST tests.
This is used to set up a very simple AST context so that AST nodes
can be allocated as usual.

No functionality change.
